refactor(pdf): extract canonical visual ordering (#226)
